requirements.txt
streamlit
Pillow
requests
beautifulsoup4